requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1

pytest==7.4.0
pytest-flask==1.3.0

factory-boy==3.3.0
Faker==24.8.0